requirements.txt
anyio==3.6.2
bcrypt==4.0.1
boto3==1.26.16
botocore==1.29.16
click==8.1.3
dnspython==2.2.1
fastapi==0.88.0
fastapi-csrf-protect==0.2.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
jmespath==1.0.1
mangum==0.17.0
motor==3.1.1
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.6
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
